// lib/form-type.ts
export interface FormData {
  // High correlation arrays
  HR: string[];
  MAP: string[];
  O2Sat: string[];
  SBP: string[];
  Resp: string[];
  // Demographics
  Unit1: string;
  Gender: string;
  HospAdmTime: string;
  Age: string;
  // Vital signs
  DBP: string;
  Temp: string;
  // Blood chemistry
  Glucose: string;
  Potassium: string;
  Hct: string;
  FiO2: string;
  Hgb: string;
  pH: string;
  BUN: string;
  WBC: string;
  Magnesium: string;
  Creatinine: string;
  Platelets: string;
  Calcium: string;
  PaCO2: string;
  BaseExcess: string;
  Chloride: string;
  HCO3: string;
  Phosphate: string;
  EtCO2: string;
  SaO2: string;
  PTT: string;
  Lactate: string;
  AST: string;
  Alkalinephos: string;
  Bilirubin_total: string;
  TroponinI: string;
  Fibrinogen: string;
  Bilirubin_direct: string;
}